% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_test.R
\name{do_test_i}
\alias{do_test_i}
\title{Title In Title Case}
\usage{
do_test_i(
  out_i,
  test_fun,
  test_args = list(0),
  fit_name = "fit",
  mc_out_name = "mc_out",
  results_fun = NULL,
  results_args = list()
)
}
\value{
Specify what are returned.
}
\description{
One paragraph description.
}
\details{
Details
(Include subjects for verbs.)
(Use 3rd person forms for verbs.)
}
\examples{
\donttest{
}

}
