% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_data.R
\name{gen_data}
\alias{gen_data}
\title{Title In Title Case}
\usage{
gen_data(
  nrep = 10,
  model,
  pop_es,
  n = 100,
  seed = NULL,
  number_of_indicators = NULL,
  reliability = NULL,
  parallel = FALSE,
  progress = FALSE,
  ncores = NULL
)
}
\value{
Specify what are returned.
}
\description{
One paragraph description.
}
\details{
Details
(Include subjects for verbs.)
(Use 3rd person forms for verbs.)
}
\examples{
\donttest{
}

}
