% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fit.R
\name{get_fit}
\alias{get_fit}
\title{Get a Fit Object From a 'power4test' Object}
\usage{
get_fit(
  object,
  which = "fit",
  fit_class = c("lavaan", "lm_list_lmhelprs"),
  i = 1
)
}
\arguments{
\item{object}{A \code{power4test} object,
such as the output of \code{\link[=power4test]{power4test()}}.}

\item{which}{The name of the fit object
to be retrieved. If set to \code{NULL},
it returns the names of supported
fit objects.}

\item{fit_class}{A character vector
of the classes
of fit objects to be retrieved.}

\item{i}{The replication from which
the fit object is to be retrieved.}
}
\value{
If a specific object is requested,
it returns the fit object, such as
the output of \code{\link[lavaan:sem]{lavaan::sem()}} or
\code{\link[lmhelprs:many_lm]{lmhelprs::many_lm()}}.

If \code{which} is set to \code{NULL}, then
it returns a character vector of
the names of the supported fit objects.
}
\description{
A helper to get the
fit object (e.g., a \code{lavaan} output)
for a replication the output of
\code{\link[=power4test]{power4test()}} and friends.
}
\details{
There are cases in which users would like
to examine the fit results in a replication.
If the sample size of a replication is
large enough, the fit results can also
be used to check the specification of
the model. The helper \code{\link[manymome:get_fit]{get_fit()}} is
for extracting the stored fit results
from the output of \code{\link[=power4test]{power4test()}}
and friends.
}
\examples{

# Specify the model

model_simple_med <-
"
m ~ x
y ~ m + x
"

# Specify the population values

model_simple_med_es <-
"
m ~ x: m
y ~ m: l
y ~ x: n
"

# Just a test with only two replications
out <- power4test(nrep = 2,
                  model = model_simple_med,
                  pop_es = model_simple_med_es,
                  n = 100,
                  test_fun = test_parameters,
                  test_args = list(pars = "m~x"),
                  iseed = 1234,
                  parallel = FALSE,
                  progress = TRUE)

get_fit(out)

}
\seealso{
See \code{\link[=power4test]{power4test()}} for
the all-in-one function, on which
this function is to be used.
}
