% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_all.R
\name{merge_all}
\alias{merge_all}
\title{Title In Title Case}
\usage{
merge_all(data_all, fit_all, mc_all = NULL)
}
\value{
Specify what are returned.
}
\description{
One paragraph description.
}
\details{
Details
(Include subjects for verbs.)
(Use 3rd person forms for verbs.)
}
\examples{
\donttest{
}

}
