% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_from_power_plot.R
\name{plot.n_from_power}
\alias{plot.n_from_power}
\title{Plot The Results of 'n_from_power'}
\usage{
\method{plot}{n_from_power}(
  x,
  what = c("ci", "power_curve", "final_n", "final_power", "target_power"),
  text_what = c("final_n", "final_power"),
  digits = 3,
  main = paste0("Power Curve ", "(Target Power: ", formatC(x$target_power, digits =
    digits, format = "f"), ")"),
  xlab = "Sample Size",
  ylab = "Estimated Power",
  pars_ci = list(),
  pars_ci_final_sample_size = list(lwd = 2, length = 0.2, col = "blue"),
  pars_target_power = list(lty = "dotted"),
  pars_final_sample_size = list(lty = "dotted"),
  pars_final_power = list(lty = "dotted", col = "blue"),
  pars_text_final_sample_size = list(y = 0, pos = 3, cex = 1),
  pars_text_final_power = list(pos = 3, cex = 1),
  ...
)
}
\arguments{
\item{x}{A \code{n_from_power} object,
the output of \code{\link[=n_from_power]{n_from_power()}}.}

\item{what}{A character vector of
what to include in the
plot. Possible values are
\code{"ci"} (confidence intervals
for the estimated sample size),
\code{"power_curve"} (the crude power
curve, if available), \code{"final_n"}
(a vertical line for
the sample size with estimated
power close enough to the target
power by confidence interval),
\code{"final_power"} (a horizontal line
for the estimated power of the
final sample size), and
\code{"target_power"} (a horizontal
line for the target power).
By default, all these elements will
be plotted.}

\item{text_what}{A character vector
of what numbers to be added as
labels. Possible values are
\code{"final_n"} (the sample size with
estimated power close enough to
the target power by confidence interval)
and \code{"final_power"} (the estimated
power of the final sample sizes).
By default, all these labels will
be added.}

\item{digits}{The number of digits
after the decimal that will be used
when adding numbers.}

\item{main}{The title of the plot.}

\item{xlab, ylab}{The labels for the
horizontal and vertical axes,
respectively.}

\item{pars_ci}{A named list of
arguments to be passed to \code{\link[=arrows]{arrows()}}
to customize the drawing of the
confidence intervals.}

\item{pars_ci_final_sample_size}{A named list of
arguments to be passed to \code{\link[=arrows]{arrows()}}
to customize the drawing of the
confidence interval of the final
sample size.}

\item{pars_target_power}{A named list
of arguments to be passed to \code{\link[=abline]{abline()}}
when drawing the horizontal line
for the target power.}

\item{pars_final_sample_size}{A
named list of arguments to be passed
to \code{\link[=abline]{abline()}} when drawing the
vertical line for the final sample
size.}

\item{pars_final_power}{A
named list of arguments to be passed
to \code{\link[=abline]{abline()}} when drawing the
horizontal line for the estimated
power at the final sample size.}

\item{pars_text_final_sample_size}{A
named list of arguments to be passed
to \code{\link[=text]{text()}} when adding the
label for the final sample size.}

\item{pars_text_final_power}{A
named list of arguments to be passed
to \code{\link[=text]{text()}} when adding the
label for the estimated power
of final sample size.}

\item{...}{Optional arguments.
Passed to \code{\link[=plot]{plot()}} when drawing
the estimated power against the
sample sizes.}
}
\value{
The \code{plot}-method of \code{n_from_power}
is called for its side effect.
}
\description{
It plot the results
of 'n_from_power', such as the
estimated power against sample size.
}
\details{
It currently plots only the relation
between estimated power and
sample size. More plots may be added
in the future.
}
\examples{

# TO PREPARE
x <- 1
\donttest{
}

}
\seealso{
\code{\link[=n_from_power]{n_from_power()}}
}
