% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power4test_by_pop_es.R
\name{power4test_by_pop_es}
\alias{power4test_by_pop_es}
\alias{c.power4test_by_pop_es}
\alias{get_rejection_rates_by_pop_es}
\title{Power By Effect Sizes}
\usage{
power4test_by_pop_es(
  object,
  pop_es_name = NULL,
  pop_es_values = NULL,
  progress = TRUE,
  ...,
  by_seed = NULL
)

\method{c}{power4test_by_pop_es}(..., sort = TRUE)

get_rejection_rates_by_pop_es(object_by_es, all_columns = FALSE)
}
\arguments{
\item{object}{A \code{power4test} object,
or a \code{power4test_by_pop_es} object.
If it is a \code{power4test_by_pop_es} object,
the first element will be used
for running the simulation.}

\item{pop_es_name}{The name of the
parameter. See the help page
of \code{\link[=ptable_pop]{ptable_pop()}} on the names for
the argument \code{pop_es}.}

\item{pop_es_values}{A numeric
vector of the population values
of the parameter specified in
\code{pop_es_names}.}

\item{progress}{Logical. Whether
progress of the simulation will be
displayed.}

\item{...}{For \code{\link[=power4test_by_pop_es]{power4test_by_pop_es()}},
they are arguments to be passed
to \code{\link[=power4test]{power4test()}}. For \code{\link[=power4test_by_pop_es]{power4test_by_pop_es()}},
they are \code{\link[=power4test_by_pop_es]{power4test_by_pop_es()}} outputs
to be combined together.}

\item{by_seed}{If set to a number,
it will be used to generate the
seeds for each call to \code{\link[=power4test]{power4test()}}.
If \code{NULL}, the default, then seeds
will still be randomly generated but
the results cannot be easily reproduced.}

\item{sort}{WHen combining \code{power4test_by_pop_es}
objects, whether they will be sorted
by effect size. Default is \code{TRUE}.}

\item{object_by_es}{A \code{power4test_by_pop_es}
object, which is an output of
\code{\link[=power4test_by_pop_es]{power4test_by_pop_es()}}.}

\item{all_columns}{If \code{TRUE}, all
columns stored by a test will be
printed. Default is \code{FALSE} and
only essential columns related to
power will be printed.}
}
\value{
The function
\code{\link[=power4test_by_pop_es]{power4test_by_pop_es()}} returns a
\code{power4test_by_pop_es} object, which is a
list of \code{power4test} objects, one for
each effect size.

The method \code{\link[=c.power4test_by_pop_es]{c.power4test_by_pop_es()}} returns
a \code{power4test_by_pop_es} object with
all the elements (tests for different
sample sizes) combined.

The function \code{\link[=get_rejection_rates_by_pop_es]{get_rejection_rates_by_pop_es()}}
returns a data frame which is
similar to the output of
\code{\link[=get_rejection_rates]{get_rejection_rates()}}, with a
column added for the effect size (\code{pop_es_name} and
\code{pop_es_values})
for each test.
}
\description{
Estimate power for a
range of effect sizes.

The method \code{\link[=c.power4test_by_pop_es]{c.power4test_by_pop_es()}}
is used to combine tests from different
runs of \code{\link[=power4test_by_pop_es]{power4test_by_pop_es()}}.

The function \code{\link[=get_rejection_rates_by_pop_es]{get_rejection_rates_by_pop_es()}}
is used to extract the rejection
rates form a \code{get_rejection_rates_by_pop_es}
object, with effect sizes added to
the output.
}
\details{
This function regenerates
datasets for a set of effect sizes
and does the stored tests in each of
them.

Optionally, it can also be run
on a object with no stored tests.
In this case, additional arguments
must be set to instruct \code{\link[=power4test]{power4test()}}
on the tests to be conducted.

It is usually used to examine the
power over a range of effect sizes.
}
\examples{


model_simple_med <-
"
m ~ x
y ~ m + x
"

model_simple_med_es <- c("y ~ m" = "l",
                         "m ~ x" = "m",
                         "y ~ x" = "n")

sim_only <- power4test(nrep = 2,
                       model = model_simple_med,
                       pop_es = model_simple_med_es,
                       n = 100,
                       R = 40,
                       ci_type = "boot",
                       fit_model_args = list(fit_function = "lm"),
                       do_the_test = FALSE,
                       iseed = 1234)

test_out <- power4test(object = sim_only,
                       test_fun = test_indirect_effect,
                       test_args = list(x = "x",
                                        m = "m",
                                        y = "y",
                                        boot_ci = TRUE,
                                        mc_ci = FALSE))

power_all_test_only_new_es <- power4test(object = test_out,
                                         pop_es = c("y ~ m" = ".10"))

out <- power4test_by_pop_es(test_out,
                            pop_es_name = "y ~ m",
                            pop_es_values = c(.10, .20))
out_reject <- get_rejection_rates_by_pop_es(out)
out_reject

}
\seealso{
\code{\link[=power4test]{power4test()}}
}
