% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_curve.R
\name{power_curve_x}
\alias{power_curve_x}
\alias{print.power_curve}
\alias{power_curve_by_n}
\alias{power_curve_by_pop_es}
\title{Power Curve}
\usage{
power_curve_x(
  object,
  formula = NULL,
  start = NULL,
  lower_bound = NULL,
  upper_bound = NULL,
  nls_args = list(),
  nls_control = list(),
  verbose = FALSE
)

\method{print}{power_curve}(x, data_used = FALSE, digits = 3, right = FALSE, row.names = FALSE, ...)

power_curve_by_n(
  object,
  formula = reject ~ I((x - c0)^e)/(b + I((x - c0)^e)),
  start = c(b = 2, c0 = 100, e = 1),
  lower_bound = c(b = 0, c0 = 0, e = 1),
  upper_bound = c(b = Inf, c0 = Inf, e = Inf),
  nls_args = list(),
  nls_control = list(),
  verbose = TRUE
)

power_curve_by_pop_es(
  object,
  formula = list(reject ~ 1 - 1/I((1 + (x/d)^a)^b), reject ~ 1 - exp(x/a)/I((1 +
    exp(x/a))^b), reject ~ 1 - 2/(exp(x/d) + exp(-x/d)), reject ~ 1/(1 + a * exp(-b *
    x))),
  start = list(c(a = 2, b = 4, d = 4), c(a = 1, b = 2), c(d = 1), c(a = 1, b = 1)),
  lower_bound = NULL,
  upper_bound = NULL,
  nls_args = list(),
  nls_control = list(),
  verbose = TRUE
)
}
\arguments{
\item{object}{An object of the class
\code{power4test_by_n} or \code{power4test_by_pop_es}.}

\item{formula}{A formula of the model
for \code{\link[stats:nls]{stats::nls()}}. It can also be
a list of formulas, and the models
will be fitted successively by
\code{\link[stats:nls]{stats::nls()}}, with the first
model fitted successfully adopted.
The response variable must be named
\code{reject}, and the predictor named
\code{x}. Whether \code{x} represents \code{n} or
\code{es} depends on the class of \code{object}.}

\item{start}{Either a named vector
of the start value(s) of parameter(s)
in \code{formula}, or a list of named
vectors of the starting value(s)
of the list of formula(s).}

\item{lower_bound}{Either a named vector
of the lower bound(s) of parameter(s)
in \code{formula}, or a list of named
vectors of the lower bound(s)
for the list of formula(s). They will
be passed to \code{lower} of \code{\link[stats:nls]{stats::nls()}}.}

\item{upper_bound}{Either a named vector
of the upper bound(s) of parameter(s)
in \code{formula}, or a list of named
vectors of the upper bound(s)
for the list of formula(s). They will
be passed to \code{upper} of \code{\link[stats:nls]{stats::nls()}}.}

\item{nls_args}{A named list of
arguments to be used when calling
\code{\link[stats:nls]{stats::nls()}}. Used to override
internal default, such as the
algorithm (default is \code{"port"}).
Used with cautions.}

\item{nls_control}{A named list of
arguments to be passed the \code{control}
argument of \code{\link[stats:nls]{stats::nls()}} when
estimating the relation between
power and sample size. The values will
override internal default values,
and also override \code{nls_args}.}

\item{verbose}{Logical. Whether
the messages will be printed when
trying different models.}

\item{x}{A \code{power_curve} object.}

\item{data_used}{Logical. Whether
the rejection rates data frame
used to fit the model is printed.}

\item{digits, right, row.names}{Arguments of the same names used
by the \code{print} method of a
\code{data.frame} object. Used when \code{data_used}
is \code{TRUE} and the rejection rates
data frame is printed.}

\item{...}{For the \code{print} method of
\code{power_curve} objects, they are optional
arguments to be passed to
\code{\link[=print.data.frame]{print.data.frame()}} when printing
the rejection rates data frame.}
}
\value{
TODO: Specify what are returned.
It returns a list which is
a \code{power_curve} object, with the
following elements:

The \code{print} method of \code{power_curve}
object returns \code{x} invisibly. Called
for its sideeffect.
}
\description{
Estimate the relation
between power and a characteristic,
such as sample size or population
effect size.
}
\details{
This is a general function,
to be used by wrappers such as
\code{\link[=power_curve_by_n]{power_curve_by_n()}} and
\code{\link[=power_curve_by_pop_es]{power_curve_by_pop_es()}}.
}
\examples{
x <- 1
# TODO: Add some examples.
\donttest{
}

}
\seealso{
\code{\link[=power4test_by_n]{power4test_by_n()}} and \code{\link[=power4test_by_pop_es]{power4test_by_pop_es()}}
}
