% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rejection_rates.R
\name{rejection_rates}
\alias{rejection_rates}
\alias{rejection_rates.default}
\alias{rejection_rates.power4test}
\alias{rejection_rates.power4test_by_es}
\alias{rejection_rates.power4test_by_n}
\title{Rejection Rates}
\usage{
rejection_rates(object, ...)

\method{rejection_rates}{default}(object, ...)

\method{rejection_rates}{power4test}(
  object,
  all_columns = FALSE,
  ci = TRUE,
  level = 0.95,
  se = FALSE,
  ...
)

\method{rejection_rates}{power4test_by_es}(object, all_columns = FALSE, ...)

\method{rejection_rates}{power4test_by_n}(object, all_columns = FALSE, ...)
}
\arguments{
\item{object}{The object
from which the rejection rates
are to be extracted, such as
a \code{power4test} object,
a \code{power4test_by_n} object,
or a \code{power4test_by_es} object.}

\item{...}{Optional arguments. Not
used.}

\item{all_columns}{If \code{TRUE}, all
columns stored by a test will be
printed. Default is \code{FALSE} and
only essential columns related to
power will be printed.}

\item{ci}{If \code{TRUE}, confidence
intervals for the rejection rates
(column \code{reject} or \code{sig}) will
be computed. Normal approximation
is used.}

\item{level}{The level of confidence
for the confidence intervals, if
\code{ci} is \code{TRUE}.}

\item{se}{If \code{TRUE}, standard errors
for the rejection rates
(column \code{reject} or \code{sig}) will
be computed. Normal approximation
is used.}
}
\value{
The \code{rejection_rates} method for
\code{power4test} objects returns
a data frame with the number of
rows equal to the number of tests.
Note that some tests, such as
the test by \code{\link[=test_parameters]{test_parameters()}},
conduct one test for each parameters.
Each such test is counted as one
test.
The data frame has these columns:
\itemize{
\item \code{test}: The name of the test.
\item \code{label}: The label for each
test, or \code{"Test"} if a test only
does on test (e.g., \code{\link[=test_indirect_effect]{test_indirect_effect()}}).
\item \code{pvalid}: The proportion of valid
tests across all replications.
\item \code{reject}: The rejection rate for
each test. If the null hypothesis
is false, then this is the power.
}

The \code{rejection_rates} method for
for \code{power4test_by_es} objects
returns a data frame which is
similar to the output of
\code{\link[=rejection_rates]{rejection_rates()}}, with a
column added for the effect size (\code{pop_es_name} and
\code{pop_es_values})
for each test.

The \code{rejection_rates} method for
for \code{power4test_by_n} objects
returns a data frame which is
similar to the output of
for a \code{power4test} object, with a
column \code{n} added for the sample size
for each test.
}
\description{
Get all rejection rates
of all tests stored in a \code{power4test}
object or other supported objects.
}
\details{
For a \code{power4test} object,
it loops over the tests stored
in a \code{power4test} object and retrieve
the rejection rate of each test.

The \code{rejection_rates} method for
\code{power4test_by_es} objects
is used to extract the rejection
rates from a \code{power4test_by_es}
object, with effect sizes added to
the output.

The \code{rejection_rates} method for
\code{power4test_by_n} objects
is used to extract the rejection
rates, with sample sizes added to
the output.
}
\examples{

model_simple_med <-
"
m ~ x
y ~ m + x
"

model_simple_med_es <- c("y ~ m" = "l",
                         "m ~ x" = "m",
                         "y ~ x" = "n")

sim_only <- power4test(nrep = 4,
                       model = model_simple_med,
                       pop_es = model_simple_med_es,
                       n = 100,
                       R = 50,
                       ci_type = "boot",
                       fit_model_args = list(fit_function = "lm"),
                       do_the_test = FALSE,
                       iseed = 1234)

test_out <- power4test(object = sim_only,
                       test_fun = test_indirect_effect,
                       test_args = list(x = "x",
                                        m = "m",
                                        y = "y",
                                        boot_ci = TRUE,
                                        mc_ci = FALSE))
test_out <- power4test(object = test_out,
                       test_fun = test_parameters)
rejection_rates(test_out)

# See the help pages of power4test_by_n() and power4test_by_es()
# for other examples.

}
\seealso{
\code{\link[=power4test]{power4test()}},
\code{\link[=power4test_by_n]{power4test_by_n()}}, and
\code{\link[=power4test_by_es]{power4test_by_es()}}.
}
