% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rig_rs.R
\name{rig_rs}
\alias{rig_rs}
\title{Random Variables Using the IG method.}
\usage{
rig_rs(
  n = 10,
  sigma = diag(1),
  skew = 0,
  kurt = 0,
  rIG_args = list(),
  pmean = 0,
  psd = 1
)
}
\arguments{
\item{n}{The number of cases
to generate.}

\item{sigma}{The population
covariance matrix. It is
recommended to supply a correlation matrix
(with diagonal elements equal to one).
The numbers will be rescaled to have
population means and SDs specified
by \code{pmean} and \code{psd}, respectively.}

\item{skew}{A numeric vector of the
population skewness coefficients.
If it is a scalar,
its value will be used as the population
skewness of all variables to be generated.}

\item{kurt}{A numeric vector of the
population excess kurtosis coefficients.
If it is a scalar,
its value will be used as the population
excess kurtosis coefficients. of all
variables to be generated.}

\item{rIG_args}{A list of extra
arguments to be passed to
\code{\link[covsim:rIG]{covsim::rIG()}}.}

\item{pmean}{A numeric vector
population means. If it is a scalar,
its value will be used as the population
means of all variables to be generated.}

\item{psd}{A numeric vector of
population standard
deviations. If it is a scalar,
its value will be used as the population
standard deviations of all variables
to be generated.}
}
\value{
A two-dimension matrix of the
generated
random numbers.
}
\description{
Generate multivariate
random numbers
using the IG method.
}
\details{
First, random numbers will be generated
using \code{\link[covsim:rIG]{covsim::rIG()}}, with the specified
target population marginal skewness and
excess kurtosis.
The random numbers will then be
rescaled with the desired population
means and standard deviations.
}
\examples{
set.seed(90870962)
sigma <- matrix(c(1, .3, .3, 1), 2, 2)
sigma
x <- rig_rs(n = 5000,
            sigma = sigma,
            skew = 2,
            kurt = 6
)
colMeans(x)
apply(x, 2, sd)
hist(x[, 1])

}
\references{
Foldnes, N. and Olson, U. H. (2016).
A simple simulation technique for
nonnormal data with prespecified
skewness, kurtosis, and covariance
matrix.
\emph{Multivariate Behavioral Research},
\emph{51}(2--3), 207--219.
\doi{10.1080/00273171.2015.1133274}
}
