% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_from_power_summary.R
\name{summary.n_from_power}
\alias{summary.n_from_power}
\alias{print.summary.n_from_power}
\title{Summarize 'n_from_power' Results}
\usage{
\method{summary}{n_from_power}(object, ...)

\method{print}{summary.n_from_power}(x, digits = 3, ...)
}
\arguments{
\item{object}{An \code{n_from_power}-class
object, such as the output of
\code{\link[=n_from_power]{n_from_power()}}.}

\item{...}{Additional arguments.
Not used for now.}

\item{x}{The output of
\code{\link[=summary.n_from_power]{summary.n_from_power()}}, the
\code{summary} method of
an \code{n_from_power} object,
which is the output of
\code{\link[=n_from_power]{n_from_power()}}.}

\item{digits}{The number of digits
after the decimal when printing
the results.}
}
\value{
Return an object of the class
\code{summary.n_from_power}, which is
simply the output of \code{\link[=n_from_power]{n_from_power()}},
with a \code{print} method dedicated for
detailed summary.

The \code{print}-method of \code{summary.n_from_power}
objects returns the object \code{x}
invisibly.
It is called for its side effect.
}
\description{
The summary method of
the output of \code{\link[=n_from_power]{n_from_power()}}.
}
\details{
It simply prepare the
results of \code{\link[=n_from_power]{n_from_power()}}
to be printed in details.
}
\examples{

# TO PREPARE
x <- 1
\donttest{
}

}
\seealso{
\code{\link[=n_from_power]{n_from_power()}}
}
