% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_fit_measure.R
\name{test_fit_measure}
\alias{test_fit_measure}
\title{Test Model Fit}
\usage{
test_fit_measure(
  fit = fit,
  model_to_fit = NULL,
  fit_measure = "chisq",
  sig_value = c(chisq = "pvalue", chisq.scaled = "pvalue.scaled", rmsea = "rmsea.pvalue",
    rmsea.scaled = "rmsea.pvalue.scaled", rmsea.robust = "rmsea.pvalue.robust"),
  sig_if = "<.05",
  check_post_check = TRUE,
  refit_args = list(se = "none"),
  always_refit = FALSE,
  fitmeasures_args = list(),
  override_measurement_model = FALSE,
  model_measurement = NULL,
  fit_name = "fit",
  get_map_names = FALSE,
  get_test_name = FALSE
)
}
\arguments{
\item{fit}{The fit object. Must be
the output of \code{\link[lavaan:lavaan]{lavaan::lavaan()}} or
its wrappers, such as \code{\link[lavaan:sem]{lavaan::sem()}}
and \code{\link[lavaan:cfa]{lavaan::cfa()}}.}

\item{model_to_fit}{The model to be
fitted, specified by \code{lavaan} model
syntax. Can contain only the structural
part, with the measurement part, if any,
retrieved from the data generation
model.
If \code{NULL}, then the model
used by \code{fit_model()} will be used.}

\item{fit_measure}{The name of the
fit measure to be used to do the test.
Must be a name that can be accepted by
\code{\link[lavaan:fitMeasures]{lavaan::fitMeasures()}}. It can also
be the test statistic, such as the model
chi-square.}

\item{sig_value}{The name of the element
of the output of \code{\link[lavaan:fitMeasures]{lavaan::fitMeasures()}}
to be used to do the test.
Must be a name that can be accepted by
\code{\link[lavaan:fitMeasures]{lavaan::fitMeasures()}}. Used when
the value used to do the test (e.g., a p-value)
is different from the value specified
in \code{fit_measure} (e.g., \code{"pvalue"} is
the p-value for model chi-square, \code{"chisq"}).
It can be
a named character vector, with the names
being possible values for \code{fit_measure}.
The name to be used will then be retrieved
based on \code{fit_measure}.}

\item{sig_if}{The criterion for doing
the test, as a character string.
For example, it is \code{"<.05"} if the test
is "significant" when the p-value is less
than .05. It is \code{"<.95"} if the "test"
is "significant" when the fit measure
(e.g., CFI) is less than .95. It must
be a string that, if appended to the
value retrieved by \code{sig_value}, is
an expression that can be evaluated
in R (e.g., ".023<.05").}

\item{check_post_check}{Logical. If
\code{TRUE}, the default, and the model
is fitted by \code{lavaan}, the test
will be conducted only if the model
passes the \code{post.check} conducted
by \code{\link[lavaan:lavInspect]{lavaan::lavInspect()}} (with
\code{what = "post.check"}).}

\item{refit_args}{A named list
of arguments to be passed to \code{\link[=fit_model]{fit_model()}}
if \code{model_to_fit} is set. If \code{model_to_fit}
is \code{NULL} and \code{always_refit} is \code{TRUE},
this list of argument values will
be used when calling \code{\link[lavaan:sem]{lavaan::sem()}},
overriding values stored, if any.}

\item{always_refit}{Whether the model
will always be fitted again. If \code{TRUE}
and \code{model_to_fit} is \code{NULL}, then the
stored model will be fitted again,
but with \code{refit_args} used. Ignored
if \code{model_to_fit} is explicitly set to
a \code{lavaan} model because this model
will always be fitted to the data. Useful
when the same stored model is fitted
but with different argument values.}

\item{fitmeasures_args}{A named list
of arguments to be passed to \code{\link[lavaan:fitMeasures]{lavaan::fitMeasures()}}.}

\item{override_measurement_model}{Whether \code{model_to_fit} already has
the measurement part and so will
override the stored measurement part
of the model, if any. If \code{FALSE},
\code{model_to_fit} can only specify the
structural part of the model.}

\item{model_measurement}{The model
syntax for the measurement model.
Ignored because its value will be
determined by \code{\link[=do_test]{do_test()}}.}

\item{fit_name}{The name of the
model fit object to be extracted.
Default is \code{"fit"}. Used only when
more than one model is fitted in
each replication. This should be
the name of the model on which the
test is to be conducted.}

\item{get_map_names}{Logical. Used
by \code{\link[=power4test]{power4test()}} to determine how
to extract stored information and
assign them to this function. Users
should not use this argument.}

\item{get_test_name}{Logical. Used
by \code{\link[=power4test]{power4test()}} to get the default
name of this test. Users should not
use this argument.}
}
\value{
In its normal usage, it returns
a one-row data frame with the
following columns:
\itemize{
\item \code{est}: The fit measure used for the
"test", such as the model chi-square
or the CFI.
\item \code{cilo} and \code{cihi}: \code{NA}. Not used.
\item \code{sig}: Whether the "test" is significant.
That is, whether the criterion
is met (e.g., the \emph{p}-value of the model
chi-square is less than .05, or the
CFI is less than .90).
\item \code{test_label}: An automatically generated
label for the test.
}
}
\description{
"Test" the model fit
of a model.
}
\details{
This function is to be used in
\code{\link[=power4test]{power4test()}} for "testing"
the model fit of a model
, by
setting it to the \code{test_fun}
argument.
}
\section{What "Test" Means for This Function}{
The term "test" is used in this function
merely to be consistent with other
test functions. What this function does
is to check whether a certain numeric
criterion based on a fit measure is met.

If the fit measure is the model chi-square,
then this is a test in the conventional
sense, using the \emph{p}-value of the chi-square.
Similarly, the test of close fit using
RMSEA is also a test.

However, the fit measure can also be
a descriptive measure such as CFI or TLI.
A descriptive measure is not used to "test"
the goodness of fit of a model. Nevertheless,
we can still estimate the probability
that this measure meets a criterion
(e.g., CFI less than .90). The function
\code{\link[=test_fit_measure]{test_fit_measure()}} can be used for
this purpose. The empirical "rejection rate" is
then the proportion of replications
with this fit measure meeting the criterion.
}

\section{Typical Scenarios}{
\subsection{The Fit Measure "Test" for the Fitted Model}{

When used with \code{\link[=power4test]{power4test()}}, this function
can be used to estimate the "rejection rate"
of a criterion (e.g., the \emph{p}-value of the
model chi-square less than .05, or the CFI
less than .90) for the model fitted when
calling \code{\link[=power4test]{power4test()}}.
}

\subsection{The Fit Measure "Test" for an Alternative Model}{

This function can also be used to estimate
the "rejection rate" of a criterion when
a model different from the stored fitted model.
For example, the data generation model is
a simple mediation model with a non-nil direct path
from the independent variable to the outcome.
However, we want to estimate the rejection
rate using "CFI<.90" when a complete mediation
model (the direct path is fixed to zero) is
fitted. This can be done by setting this
model to the argument \code{model_to_fit}.
}
}

\examples{

# Specify the model

mod <-
"
m ~ x
y ~ m + x
"

# Specify the population values

mod_es <-
"
y ~ m: l
m ~ x: s
y ~ x: m
"

# Simulate the data

sim_only <- power4test(
  nrep = 2,
  model = mod,
  pop_es = mod_es,
  n = 100,
  iseed = 1234
)

# Do the tests in each replication

mod_complete <-
"
m ~ x
y ~ m
"

test_out <- power4test(
  object = sim_only,
  test_fun = test_fit_measure,
  test_args = list(
    model_to_fit = mod_complete,
    fit_measure = "cfi",
    sig_if = "<.90"
  )
)

print(test_out,
      test_long = TRUE)


}
\seealso{
\code{\link[=power4test]{power4test()}}
}
