% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_3_in_1.R
\name{common_processes}
\alias{common_processes}
\title{Common Data Processing}
\usage{
common_processes(
  data,
  cut_patterns = NULL,
  cuts = NULL,
  missing_values_args = list(),
  prop = NULL,
  mech = "MCAR",
  method = c("none", "mean", "sum"),
  na.rm = FALSE
)
}
\arguments{
\item{data}{A data frame.}

\item{cut_patterns}{A named vector.
The names are the names of the latent
variables for which indicator scores
will be converted. Each value must be
the name of one of the built-in
patterns (call \code{cut_patterns()} to
list the patterns and their names).
Can be used with \code{cuts} but a latent
variable should appear only either
in \code{cut_patterns} or \code{cuts}, not both.
If both \code{cut_patterns} and \code{cuts} are set to \code{NULL},
then the original data (\code{data}) will
be returned unchanged.}

\item{cuts}{A named list.
The names are the names of the latent
variables for which indicator scores
will be converted.
Each element is a vector of the
thresholds for the conversion. \code{-Inf}
and \code{Inf} will be automatically
included during the conversion.
Can be used with \code{cut_patterns} but a latent
variable should appear only either
in \code{cut_patterns} or \code{cuts}, not both.
If both \code{cut_patterns} and \code{cuts} are set to \code{NULL},
then the original data (\code{data}) will
be returned unchanged.}

\item{missing_values_args}{A named
list of optional arguments to be
passed to \code{\link[mice:ampute]{mice::ampute()}}. Note that
\code{prop} and \code{mech} will override
the values set for them in
\code{missing_values_args}, if any.}

\item{prop}{The proportion of missingness.
Default is 0.5, about 50\% of the cases
have missing data.
If set to \code{NULL},
then the original data (\code{data}) will
be returned unchanged.}

\item{mech}{The missing data mechanism.
Default is \code{"MCAR"} (missing completely at random).
Other possible values are \code{"MAR"}
(missing at random) and \code{"MNAR"}
(missing not at random). Please refer
to the help of \code{\link[mice:ampute]{mice::ampute()}} for
details.}

\item{method}{The method to be used
to compute the scale scores. Can
be \code{"mean"} or \code{"sum"}. Implemented
by \code{\link[=mean]{mean()}} and \code{\link[=sum]{sum()}}.
If set to \code{"none"},
then the original data (\code{data}) will
be returned unchanged.}

\item{na.rm}{How missing value (\code{NA})
are handled. Default is \code{FALSE},
the same default value for \code{\link[=mean]{mean()}}
and \code{\link[=sum]{sum()}}.}
}
\value{
The function returns a data frame with
the processed data.
}
\description{
For the \code{process_data}
argument. It do three types of processing
in one step: Creating ordinal variables,
generating missing values, and computing
scales scores.
}
\details{
This function is to be used in
the \code{process_data} argument of
\code{\link[=power4test]{power4test()}}.

This function is simply a wrapper
of the following three functions:
\itemize{
\item \code{\link[=ordinal_variables]{ordinal_variables()}}
\item \code{\link[=missing_values]{missing_values()}}
\item \code{\link[=scale_scores]{scale_scores()}}
}

Please refer to these functions on
how the raw data is processed.
}
\examples{

# Specify the model

mod <-
"
m ~ x
y ~ m + x
"

# Specify the population values

mod_es <-
"
y ~ m: l
m ~ x: m
y ~ x: n
"

# Specify the numbers of indicators and reliability coefficients

k <- c(y = 3,
       m = 4,
       x = 5)
rel <- c(y = .70,
         m = .70,
         x = .70)

# Simulate the data

out <- power4test(
         nrep = 2,
         model = mod,
         pop_es = mod_es,
         n = 200,
         number_of_indicators = k,
         reliability = rel,
         process_data = list(
           fun = common_processes,
           args = list(
                     prop = .75,
                     cut_patterns = c(x = "-ma3", y = "-ma3", m = "-ma3")
                   )
           ),
         test_fun = test_parameters,
         test_args = list(op = "~"),
         parallel = FALSE,
         iseed = 1234)

dat <- pool_sim_data(out)
head(dat, 50)

}
\seealso{
\code{\link[=power4test]{power4test()}}. See also
\code{\link[=ordinal_variables]{ordinal_variables()}}, \link{missing_values},
and \link{scale_scores} for the processes
employed.
}
