% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_test.R
\name{do_test}
\alias{do_test}
\title{Title In Title Case}
\usage{
do_test(
  out_all,
  test_fun,
  test_args = list(),
  fit_name = "fit",
  mc_out_name = "mc_out",
  results_fun = NULL,
  results_args = list(),
  parallel = FALSE,
  progress = FALSE,
  ncores = max(1, parallel::detectCores(logical = FALSE) - 1)
)
}
\value{
Specify what are returned.
}
\description{
One paragraph description.
}
\details{
Details
(Include subjects for verbs.)
(Use 3rd person forms for verbs.)
}
\examples{
\donttest{
}

}
