% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{fit_model}
\alias{fit_model}
\title{Fit a Model to a List of Datasets}
\usage{
fit_model(
  sim_data_all,
  ...,
  parallel = FALSE,
  progress = FALSE,
  ncores = max(1, parallel::detectCores(logical = FALSE) - 1)
)
}
\value{
An object of the class \code{fit_model},
which is a list of the output of
\code{\link[lavaan:sem]{lavaan::sem()}}.
}
\description{
Get the output of
\code{\link[=sim_data]{sim_data()}} and fit a model to each
of the stored dataset.
}
\details{
By default, it extracts the model
stored in the output of \code{\link[=sim_data]{sim_data()}},
fits the model to each dataset
simulated using \code{\link[lavaan:sem]{lavaan::sem()}}, and
returns the results.
}
\examples{
\donttest{
}

}
