% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_mc.R
\name{gen_mc}
\alias{gen_mc}
\title{Generate Monte Carlo Estimates}
\usage{
gen_mc(
  fit_all,
  R = 100,
  seed = NULL,
  ...,
  parallel = FALSE,
  progress = FALSE,
  ncores = max(1, parallel::detectCores(logical = FALSE) - 1)
)
}
\value{
An \code{mc_list} object, which is a list
of the output of \code{\link[manymome:do_mc]{manymome::do_mc()}}.
}
\description{
Get a list of the output
of \code{\link[lavaan:sem]{lavaan::sem()}} and generate
Monte Carlo estimates of model
parameters.
}
\details{
It simply call \code{\link[manymome:do_mc]{manymome::do_mc()}}
on each of the fit output of
\code{\link[lavaan:sem]{lavaan::sem()}}. The simulated
estimates can then be used to test
effects such as indirect effects.
}
\examples{
\donttest{
}

}
