% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power4mome-package.R
\docType{package}
\name{power4mome-package}
\alias{power4mome}
\alias{power4mome-package}
\title{power4mome: Power Analysis for Moderation and Mediation}
\description{
Power analysis and sample size determination for moderation, mediation, and moderated mediation in models fitted by structural equation modeling.
}
\author{
\strong{Maintainer}: Shu Fai Cheung \email{shufai.cheung@gmail.com} (\href{https://orcid.org/0000-0002-9871-9448}{ORCID})

}
\keyword{internal}
