% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power4mome-package.R
\docType{package}
\name{power4mome-package}
\alias{power4mome}
\alias{power4mome-package}
\title{power4mome: Power Analysis for Moderation and Mediation}
\description{
Power analysis and sample size determination for moderation, mediation, and moderated mediation in models fitted by structural equation modeling.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://sfcheung.github.io/power4mome/}
  \item Report bugs at \url{https://github.com/sfcheung/power4mome/issues}
}

}
\author{
\strong{Maintainer}: Shu Fai Cheung \email{shufai.cheung@gmail.com} (\href{https://orcid.org/0000-0002-9871-9448}{ORCID})

Authors:
\itemize{
  \item Sing-Hang Cheung (\href{https://orcid.org/0000-0001-5182-0752}{ORCID})
  \item Wendie Yang (\href{https://orcid.org/0009-0000-8388-6481}{ORCID})
}

}
\keyword{internal}
