% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_curve_predict.R
\name{predict.power_curve}
\alias{predict.power_curve}
\title{Predict Method for a 'power_curve' Object}
\usage{
\method{predict}{power_curve}(object, newdata, ...)
}
\arguments{
\item{object}{A \code{power_curve}
object.}

\item{newdata}{A data frame with
a column named \code{x}. This is a
required argument.}

\item{...}{Additional arguments.
Passed to the corresponding
\code{predict} method.}
}
\value{
It returns a numeric vector of
the predicted rejection rates.
}
\description{
Compute the predicted
values in a model fitted by
\code{\link[=power_curve_x]{power_curve_x()}}.
}
\details{
It retrieve the stored results
and call the corresponding methods.
}
\examples{
# TODO:
# - To write
x <- 1

}
\seealso{
\code{\link[=power_curve_x]{power_curve_x()}}.
}
