% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptable_pop.R
\name{ptable_pop}
\alias{ptable_pop}
\alias{model_matrices_pop}
\title{Generate the Population Model}
\usage{
ptable_pop(
  model,
  pop_es,
  es1 = c(s = 0.1, m = 0.3, l = 0.5),
  es2 = c(s = 0.05, m = 0.1, l = 0.15)
)

model_matrices_pop(x, pop_es)
}
\arguments{
\item{model}{String. The model defined
by \code{lavaan} model syntax.}

\item{pop_es}{It can a data frame
with these columns: \code{lhs}, \code{op},
\code{rhs}, and \code{pop}. The first three
columns correspond to those in a
\code{lavaan} parameter table. The column
\code{pop} stores the population values.
The column \code{es} stores the original
labels, for reference. It can also be
A named character vector. The names
are \code{lavaan} for the selected
parameters. For example, \code{m ~ x}
denotes the path from \code{x} to \code{m}. Can
specify more than one parameters. For
example, \code{y ~ m + x} denotes the two
paths from \code{m} and \code{x} to \code{y}. The
value is the label for the effect
size: \code{s} for small, \code{m} for medium,
and \code{l} for large. There are two
possible values, one set, \code{es1}, for
correlations and regression
coefficients, the other set, \code{es2},
for standardized moderation effect,
the coefficients of a product term.}

\item{es1}{A named vector to set the
values for each label of the effect
size of correlations and regression
paths.
Default is \code{c("s" = .10, "m" = .30, "l" = .50)}.
Used only if \code{par_es} is a named
vector.}

\item{es2}{A named vector to set the
values for each label of the effect
size of product term.
Default is \code{c("s" = .05, "m" = .10, "l" = .15)}.
Used only if \code{par_es} is a named
vector.}

\item{x}{It can be 'lavaan' model
syntax, to be passed to \code{\link[=ptable_pop]{ptable_pop()}},
or a parameter table with the column
\code{start} set to the population values,
such as the output of \code{\link[=ptable_pop]{ptable_pop()}}.}
}
\value{
The function \code{\link[=ptable_pop]{ptable_pop()}} returns
a \code{lavaan} parameter table of the
model, with \code{start} set to the
population values.

The function \code{\link[=model_matrices_pop]{model_matrices_pop()}}
returns a \code{lavaan} LISREL-style model
matrices (like the output of
\code{\link[lavaan:lavInspect]{lavaan::lavInspect()}} with \code{what}
set to \code{free}), with matrix elements
set to the population values. If
\code{x} is the model syntax, it will be
stored in the attributes \code{model}.
}
\description{
Generate the complete
population model using the model
syntax and user-specified effect
sizes.
}
\details{
The function \code{\link[=ptable_pop]{ptable_pop()}} generates a \code{lavaan}
parameter table that can be used
to generate data based on the population
values of model parameters.

The function \code{\link[=model_matrices_pop]{model_matrices_pop()}}
generates models matrices with
population values.
}
\section{Functions}{
\itemize{
\item \code{model_matrices_pop()}: 

}}
\examples{

model1 <-
"
m1 ~ x + c1
m2 ~ m1 + x2 + c1
y ~  m2 + m1 + x + w + x:w + c1
"

model1_es <- c("m1 ~ x" = "-m",
               "m2 ~ m1" = "s",
               "y ~ m2" = "l",
               "y ~ x" = "m",
               "y ~ w" = "s",
               "y ~ x:w" = "s",
               "x ~~ w" = "s")

ptable_final1 <- ptable_pop(model1,
                            pop_es = model1_es)
ptable_final1


model_matrices_pop(ptable_final1)

model_matrices_pop(model1,
                   pop_es = model1_es)

}
\author{
Shu Fai Cheung \url{https://orcid.org/0000-0002-9871-9448}
}
