% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptable_pop.R
\name{ptable_pop}
\alias{ptable_pop}
\alias{model_matrices_pop}
\title{Generate the Population Model}
\usage{
ptable_pop(
  model = NULL,
  pop_es = NULL,
  es1 = c(n = 0, s = 0.1, m = 0.3, l = 0.5),
  es2 = c(n = 0, s = 0.05, m = 0.1, l = 0.15),
  standardized = TRUE,
  n_std = 1e+05,
  std_force_monte_carlo = FALSE
)

model_matrices_pop(x, ..., drop_list_single_group = TRUE)
}
\arguments{
\item{model}{String. The model defined
by \code{lavaan} model syntax.}

\item{pop_es}{It can a data frame
with these columns: \code{lhs}, \code{op},
\code{rhs}, and \code{pop}. The first three
columns correspond to those in a
\code{lavaan} parameter table. The column
\code{pop} stores the population values.
The column \code{es} stores the original
labels, for reference. It can also be
A named character vector. See 'Details'
on how to specify this vector.}

\item{es1}{A named vector to set the
values for each label of the effect
size of correlations and regression
paths.
Default is \code{c("n" = .00, "s" = .10, "m" = .30, "l" = .50)}.
Used only if \code{pop_es} is a named
vector.}

\item{es2}{A named vector to set the
values for each label of the effect
size of product term.
Default is \code{c("n" = .00, "s" = .05, "m" = .10, "l" = .15)}.
Used only if \code{pop_es} is a named
vector.}

\item{standardized}{Logical. If
\code{TRUE}, the default, variances and
error variances are scaled to ensure
the population variances of the
endogenous variables are close to
one, and hence the effect sizes are
standardized effect sizes if the
variances of the continuos exogenous
variables are also equal to one.}

\item{n_std}{The sample size used to
determine the error variances by
simulation. Default is 100000.}

\item{std_force_monte_carlo}{Logical.
If \code{FALSE}, the default,
standardization is done analytically
if the model has no product terms,
and by simulation if the model has
product terms. If \code{TRUE}, simulation
will be used whether the model has
product terms or not. Always fall
back to standardization if
analytical standardization failed.}

\item{x}{It can be 'lavaan' model
syntax, to be passed to \code{\link[=ptable_pop]{ptable_pop()}},
or a parameter table with the column
\code{start} set to the population values,
such as the output of \code{\link[=ptable_pop]{ptable_pop()}}.}

\item{...}{If \code{x} is a model syntax,
these are arguments to be passed to
\code{\link[=ptable_pop]{ptable_pop()}}.}

\item{drop_list_single_group}{If
\code{TRUE} and the number groups is
equal to one, the output will be
a list of matrices of one group
only. Default if \code{TRUE}.}
}
\value{
The function \code{\link[=ptable_pop]{ptable_pop()}} returns
a \code{lavaan} parameter table of the
model, with \code{start} set to the
population values.

The function \code{\link[=model_matrices_pop]{model_matrices_pop()}}
returns a \code{lavaan} LISREL-style model
matrices (like the output of
\code{\link[lavaan:lavInspect]{lavaan::lavInspect()}} with \code{what}
set to \code{free}), with matrix elements
set to the population values. If
\code{x} is the model syntax, it will be
stored in the attributes \code{model}.
If the model is a multigroup model
with \emph{k} groups (\emph{k} greater than 1),
then it returns a list of \emph{k} lists
of \code{lavaan} LISREL-style model
matrices.
}
\description{
Generate the complete
population model using the model
syntax and user-specified effect
sizes.
}
\details{
The function \code{\link[=ptable_pop]{ptable_pop()}} generates a \code{lavaan}
parameter table that can be used
to generate data based on the population
values of model parameters.
\subsection{Setting \code{pop_es} to a named vector}{

If \code{pop_es} is specified by a named
vector, it must follow the convention
below.
\itemize{
\item The names of the vectors are
\code{lavaan} names for the selected
parameters. For example, \code{m ~ x}
denotes the path from \code{x} to \code{m}.
\item Alternatively, the names can be
either \code{".beta."} or \code{".cov."}.
Use \code{".beta."} to set the default
values for all regression coefficients.
Use \code{".cov."} to set the default
values for all correlations of
exogenous variables (e.g., predictors).
\item If using \code{lavaan} names, can
specify more than one parameters
using \code{+}. For example, \code{y ~ m + x}
denotes the two paths from \code{m} and
\code{x} to \code{y}.
\item The value of each element can be
the label for the effect size: \code{n}
for nil, \code{s} for small, \code{m} for
medium, and \code{l} for large. The
value for each label is determined
by \code{es1} and \code{es2}.
\item The value of \code{pop_es} can also be
set to a value, but it must be
quoted as a string, such as \code{"y ~ x" = ".31"}.
}

The vector \code{es1} is for correlations
and regression coefficients, and the
vector \code{es2} is for for standardized
moderation effect, the coefficients
of a product term.
}

\subsection{Multigroup Models}{

The function also support multigroup
models.

Because the model is the population
model, equality constraints are
irrelevant and the model syntax
specifies only the \emph{form} of the
model. Therefore, \code{model} is
specified as in the case of single
group models.

For \code{pop_es}, instead of using a
named vectors, use as named \emph{list}.
\itemize{
\item The names are the parameters, or
keywords such as \code{.beta.} and
\code{.cov.}, like specifying the
population values for a single
group model.
\item The elements are character vectors.
If it has only one element (e.g.,
a single string), then it is the
the population value for all groups.
If it has more than one element
(e.g., a vector of three strings),
then they are the population values
of the groups. For a model of \emph{k}
groups, each vector must have
either \emph{k} elements or one element.
}

This is an example:

\code{list("m ~ x" = "m", "y ~ m" = c("s", "m", "l"))}

In this model, the population value
of the path \code{m ~ x} is medium for
all groups, while the population
values for the path \code{y ~ m} are
small, medium, and large,
respectively.
}

The function \code{\link[=model_matrices_pop]{model_matrices_pop()}}
generates models matrices with
population values.
}
\section{Functions}{
\itemize{
\item \code{model_matrices_pop()}: 

}}
\examples{

model1 <-
"
m1 ~ x + c1
m2 ~ m1 + x2 + c1
y ~  m2 + m1 + x + w + x:w + c1
"

model1_es <- c("m1 ~ x" = "-m",
               "m2 ~ m1" = "s",
               "y ~ m2" = "l",
               "y ~ x" = "m",
               "y ~ w" = "s",
               "y ~ x:w" = "s",
               "x ~~ w" = "s")

ptable_final1 <- ptable_pop(model1,
                            pop_es = model1_es)
ptable_final1


model_matrices_pop(ptable_final1)

model_matrices_pop(model1,
                   pop_es = model1_es)

}
\author{
Shu Fai Cheung \url{https://orcid.org/0000-0002-9871-9448}
}
