% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_data.R
\name{sim_data}
\alias{sim_data}
\title{Simulate Datasets Based on a Model}
\usage{
sim_data(
  nrep = 10,
  model,
  pop_es,
  n = 100,
  iseed = NULL,
  number_of_indicators = NULL,
  reliability = NULL,
  parallel = FALSE,
  progress = FALSE,
  ncores = max(1, parallel::detectCores(logical = FALSE) - 1)
)
}
\value{
A list of the class \code{sim_data},
with length \code{nrep}. Each element
is a \code{sim_data_i} object, with
the following major elements:
\itemize{
\item ptable: A \code{lavaan} parameter
table of the model, with population
values set in the column \code{start}.
(It is the output of the internal
function \code{ptable_pop()}.)
\item mm_out: The population model
represented by model matrices
as in \code{lavaan}. (It is the output
of the internal function
\code{model_matrices_pop()}.)
\item mm_lm_out: A list of regression
model formula, one for each
endogenous variable. (It is the
output of the internal function
\code{mm_lm()}.)
\item mm_lm_dat_out: A simulated dataset
generated from the population model.
(It is the output of the internal
function \code{mm_lm_data()}).
\item model_original: The original model
syntax (i.e., the argument \code{model}).
\item model_final: A modified model
syntax if the model is a latent
variable model. Indicators are added
to the syntax.
\item fit0: The output of \code{\link[lavaan:sem]{lavaan::sem()}}
with \code{ptable} as the model and
\code{do.fit} set to \code{FALSE}. Use for
easy retrieve of information
about the model.
}
}
\description{
Get a model matrix and
effect size specification and
simulate a number of datasets,
along with other information.
}
\details{
It does two tasks:
\itemize{
\item Determine the actual population
model with population values based
on:
\itemize{
\item A model syntax for the observed
variables (for a path model)
or the latent factors (for a
latent variable model).
\item A textual specification of the
effect sizes of parameters.
\item The number of indicators for
each latent factor if the model
is a latent variable model.
\item The reliability of each latent
factor as measured by the
indicators if the model is a
latent factor model.
}
\item Generate \emph{m} simulated datasets
from the population model.
}

The simulated datasets can then be
used to fit a model, test
parameters, and estimate power.

The output is usually used by
\code{\link[=fit_model]{fit_model()}} to fit a target model,
using the population model, to each
of the dataset.
}
\examples{
\donttest{
}

}
