% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_data.R
\name{sim_data}
\alias{sim_data}
\title{Simulate Datasets Based on a Model}
\usage{
sim_data(
  nrep = 10,
  model,
  pop_es,
  n = 100,
  iseed = NULL,
  number_of_indicators = NULL,
  reliability = NULL,
  parallel = FALSE,
  progress = FALSE,
  ncores = max(1, parallel::detectCores(logical = FALSE) - 1)
)
}
\arguments{
\item{nrep}{The number of replications
to generate the simulated datasets.
Default is 10.}

\item{model}{The \code{lavaan} model
syntax of the population model.
Required.}

\item{pop_es}{The character to
specify population effect sizes.
See 'Details' on how to set the
effect sizes for this argument.
Required.}

\item{n}{The sample size for each
dataset. Default is 100.}

\item{iseed}{The seed for the random
number generator. Default is \code{NULL}
and the seed is not changed.}

\item{number_of_indicators}{A named
vector to specify the number of
indicators for each factors. See
'Details' on how to set this
argument. Default is \code{NULL} and all
variables in the model syntax are
observed variables.}

\item{reliability}{A named vector
to set the reliability coefficient
of each set of indicators. Default
is \code{NULL}.}

\item{parallel}{If \code{TRUE}, parallel
processing will be used to simulate
the datasets. Default is \code{FALSE}.}

\item{progress}{If \code{TRUE}, the progress
of data simulation will be displayed.
Default is `FALSE.}

\item{ncores}{The number of CPU
cores to use if parallel processing
is used.}
}
\value{
A list of the class \code{sim_data},
with length \code{nrep}. Each element
is a \code{sim_data_i} object, with
the following major elements:
\itemize{
\item ptable: A \code{lavaan} parameter
table of the model, with population
values set in the column \code{start}.
(It is the output of the internal
function \code{ptable_pop()}.)
\item mm_out: The population model
represented by model matrices
as in \code{lavaan}. (It is the output
of the internal function
\code{model_matrices_pop()}.)
\item mm_lm_out: A list of regression
model formula, one for each
endogenous variable. (It is the
output of the internal function
\code{mm_lm()}.)
\item mm_lm_dat_out: A simulated dataset
generated from the population model.
(It is the output of the internal
function \code{mm_lm_data()}).
\item model_original: The original model
syntax (i.e., the argument \code{model}).
\item model_final: A modified model
syntax if the model is a latent
variable model. Indicators are added
to the syntax.
\item fit0: The output of \code{\link[lavaan:sem]{lavaan::sem()}}
with \code{ptable} as the model and
\code{do.fit} set to \code{FALSE}. Use for
easy retrieve of information
about the model.
}
}
\description{
Get a model matrix and
effect size specification and
simulate a number of datasets,
along with other information.
}
\details{
It does two tasks:
\itemize{
\item Determine the actual population
model with population values based
on:
\itemize{
\item A model syntax for the observed
variables (for a path model)
or the latent factors (for a
latent variable model).
\item A textual specification of the
effect sizes of parameters.
\item The number of indicators for
each latent factor if the model
is a latent variable model.
\item The reliability of each latent
factor as measured by the
indicators if the model is a
latent factor model.
}
\item Generate \emph{m} simulated datasets
from the population model.
}

The simulated datasets can then be
used to fit a model, test
parameters, and estimate power.

The output is usually used by
\code{\link[=fit_model]{fit_model()}} to fit a target model,
using the population model, to each
of the dataset.
\subsection{Setting \code{pop_es}}{

(To prepare)
}

\subsection{Setting \code{number_of_indicators} and \code{reliability}}{

(To prepare)
}
}
\examples{
mod <-
"m ~ x
 y ~ m + x"
es <-
c("y ~ m" = "m",
  "m ~ x" = "m",
  "y ~ x" = "n")
data_all <- sim_data(nrep = 5,
                     model = mod,
                     pop_es = es,
                     n = 100,
                     iseed = 1234)

}
