% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_test.R
\name{summarize_tests}
\alias{summarize_tests}
\alias{print.test_summary_list}
\alias{print.test_summary}
\alias{print.test_out_list}
\title{Summarize Test Results}
\usage{
summarize_tests(object)

\method{print}{test_summary_list}(x, digits = 3, ...)

\method{print}{test_summary}(x, digits = 2, ...)

\method{print}{test_out_list}(x, digits = 3, test_long = FALSE, ...)
}
\arguments{
\item{object}{A \code{power4test} object
or the element \code{test_all} in
a \code{power4test} object.}

\item{x}{The object to be printed.}

\item{digits}{The numbers of digits
after the decimal when print
numeric results.}

\item{...}{Optional arguments.
Not used.}

\item{test_long}{If \code{TRUE}, a detailed
report will be printed.}
}
\value{
The function \code{\link[=summarize_tests]{summarize_tests()}} returns
a list of the class \code{test_summary_list}.
Each element contains a summary of a
test stored. The elements are of
the class \code{test_summary}, with
these elments:
\itemize{
\item \code{test_attributes}: The stored
information of a test, for printing.
\item \code{nrep}: The number of datasets
(replications).
\item \code{mean}: The means of numeric
information. For significance
tests, these are the rejection
rates.
\item \code{nvalid}: The number of non-\code{NA}
replications used to compute each
mean.
}

The \code{print} methods return \code{x} invisibly.
They are called for their side
effects.
}
\description{
Extract and summarize
test results.
}
\details{
This function is used to extract
information from each test stored
in a \code{power4test} object.

The method \code{print.test_out_list()} is
used to print the contest of a list
of test stored in a \code{power4test}
object, with the option to print
just the names of tests.
}
\examples{

mod <-
"
m ~ a*x
y ~ b*m + x
ab := a * b
"

mod_es <- c("y ~ m" = "l",
            "m ~ x" = "m",
            "y ~ x" = "n")

sim_only <- power4test(nrep = 2,
                       model = mod,
                       pop_es = mod_es,
                       n = 100,
                       do_the_test = FALSE,
                       iseed = 1234)

test_out <- power4test(object = sim_only,
                       test_fun = test_parameters)

test_out

}
\seealso{
\code{\link[=power4test]{power4test()}}
}
