% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_test.R
\name{summarize_tests}
\alias{summarize_tests}
\alias{print.test_summary_list}
\alias{print.test_summary}
\title{Summarize Test Results}
\usage{
summarize_tests(object)

\method{print}{test_summary_list}(x, ...)

\method{print}{test_summary}(x, ...)
}
\arguments{
\item{object}{A \code{power4test} object
or the element \code{test_all} in
a \code{power4test} object.}

\item{x}{A \code{test_summary_list}
object or a \code{test_summary} object,
depending on the print method used.}

\item{...}{Optional arguments.
Not used.}
}
\value{
The function \code{\link[=summarize_tests]{summarize_tests()}} returns
a list of the class \code{test_summary_list}.
Each element contains a summary of a
test stored. The elements are of
the class \code{test_summary}, with
these elments:
\itemize{
\item \code{test_attributes}: The stored
information of a test, for printing.
\item \code{nrep}: The number of datasets
(replications).
\item \code{mean}: The means of numeric
information. For significance
tests, these are the rejection
rates.
\item \code{nvalid}: The number of non-\code{NA}
replications used to compute each
mean.
}

The \code{print} methods of
\code{test_summary_list} and
\code{test_summary} return \code{x} invisibly.
They are called for their side
effects.
}
\description{
Extract and summarize
test results.
}
\details{
This function is used to extract
information from each test stored
in a \code{power4test} object.
}
\examples{

# TO PREPARE
x <- 1
\donttest{
}

}
\seealso{
\code{\link[=power4test]{power4test()}}
}
