% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_from_power_summary.R
\name{summary.n_from_power}
\alias{summary.n_from_power}
\alias{print.summary.n_from_power}
\title{Summarize 'n_from_power' Results}
\usage{
\method{summary}{n_from_power}(object, ...)

\method{print}{summary.n_from_power}(x, digits = 3, ...)
}
\arguments{
\item{object}{An \code{n_from_power}-class
object, such as the output of
\code{\link[=n_from_power]{n_from_power()}}.}

\item{...}{Additional arguments.
Not used for now.}

\item{x}{The output of
\code{\link[=summary.n_from_power]{summary.n_from_power()}}, the
\code{summary} method of
an \code{n_from_power} object,
which is the output of
\code{\link[=n_from_power]{n_from_power()}}.}

\item{digits}{The number of digits
after the decimal when printing
the results.}
}
\value{
Return an object of the class
\code{summary.n_from_power}, which is
simply the output of \code{\link[=n_from_power]{n_from_power()}},
with a \code{print} method dedicated for
detailed summary.

The \code{print}-method of \code{summary.n_from_power}
objects returns the object \code{x}
invisibly.
It is called for its side effect.
}
\description{
The summary method of
the output of \code{\link[=n_from_power]{n_from_power()}}.
}
\details{
It simply prepare the
results of \code{\link[=n_from_power]{n_from_power()}}
to be printed in details.
}
\examples{

mod <-
"
m ~ a*x
y ~ b*m + x
ab := a * b
"

mod_es <- c("y ~ m" = "l",
            "m ~ x" = "m",
            "y ~ x" = "n")

sim_only <- power4test(nrep = 10,
                       model = mod,
                       pop_es = mod_es,
                       n = 100,
                       do_the_test = FALSE,
                       iseed = 1234)

test_out <- power4test(object = sim_only,
                       test_fun = test_parameters,
                       test_args = list(pars = "ab"))

# In real analysis, to have more stable results:
# - Use a larger final_nrep (e.g., 500).
# - Use the default ns_per_trial of 3, or just remove it.

# If the default values are OK, this call is sufficient:
# power_vs_n <- n_from_power(test_out,
#                            target_power = .80,
#                            seed = 4567)
power_vs_n <- n_from_power(test_out,
                           progress = TRUE,
                           target_power = .80,
                           final_nrep = 10,
                           ns_per_trial = 1,
                           nrep_steps = 1,
                           max_trials = 1,
                           seed = 4567)
summary(power_vs_n)

}
\seealso{
\code{\link[=n_from_power]{n_from_power()}}
}
